requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
flask-login==0.6.3
flask-wtf==1.2.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
pillow==10.1.0
wtforms==3.1.1
werkzeug==2.3.7
